refactor: organize file structure for DashboardPage
