+++ client/components/page/Dashboard/index.ts
@@ -1,0 +1,1 @@
+export { default as DashboardPage } from './Dashboard.page';
