+++ client/components/shared/Form/index.ts
@@ -1,0 +1,1 @@
+export { default as PlanAccordionForm } from './PlanAccordionForm';
