+++ client/components/shared/Header/SidebarWithHeader.tsx
@@ -71,7 +71,6 @@ const SidebarWithHeader = ({ children }: { children: ReactNode }) => {
           <SidebarContent onClose={onClose} />
         </DrawerContent>
       </Drawer>
-      {/* mobilenav */}
       <MobileNav onOpen={onOpen} />
       <Box ml={{ base: 0, md: 60 }} p="4">
         {children}
