+++ client/components/shared/Header/index.ts
@@ -1,0 +1,1 @@
+export { default as SidebarWithHeader } from './SidebarWithHeader';
