+++ client/components/shared/input/ReactHookFormInput.tsx
@@ -9,7 +9,7 @@ import {
 import { Controller, UseFormReturn, FieldErrorsImpl } from 'react-hook-form';
 import PhoneInput from 'react-phone-input-2';
 import 'react-phone-input-2/lib/style.css';
-import { TInputField } from '../planAccordionForm';
+import { TInputField } from '.';
 
 type TErrorMessage = {
   input: TInputField;
