+++ client/pages/index.tsx
@@ -1,4 +1,5 @@
-import type { GetStaticProps } from 'next';
+import type { NextPage, GetStaticProps } from 'next';
+import Head from 'next/head';
 import {
   QueryProductFindMany,
   QueryProductFindManyVariables,
@@ -26,12 +27,25 @@ import client from '../apollo-client';
 import { GET_PRODUCT_FIND_MANY } from '../gql/product';
 import { TopPage } from '@/components/page/Top/index';
 
-export interface TopPageProps {
+export interface TopNextPageProps {
   products: QueryProductFindMany_productFindMany[];
   taxes: QueryTaxFindMany_taxFindMany[];
   plans: QueryPlanFindMany_planFindMany[];
   planOptions: QueryPlanOptionFindMany_planOptionFindMany[];
 }
+
+const TopNextPage: NextPage<TopNextPageProps> = (props: TopNextPageProps) => {
+  return (
+    <>
+      <Head>
+        <title>Top Page</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <TopPage {...props} />
+    </>
+  );
+};
 
 export const getStaticProps: GetStaticProps = async () => {
   const {
@@ -67,4 +81,4 @@ export const getStaticProps: GetStaticProps = async () => {
   };
 };
 
-export default TopPage;
+export default TopNextPage;
